requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
anyio==4.7.0
